feat(app/interface): refactor WarningBar
